[#3] Move base class to util file
